+++ floatingip_assign.go
@@ -16,7 +16,7 @@ func newFloatingIPAssignCommand(cli *CLI) *cobra.Command {
 		Args:                  cobra.ExactArgs(2),
 		TraverseChildren:      true,
 		DisableFlagsInUseLine: true,
-		PreRunE:               cli.ensureActiveContext,
+		PreRunE:               cli.ensureToken,
 		RunE:                  cli.wrap(runFloatingIPAssign),
 	}
 	cmd.MarkFlagRequired("server")