+++ floatingip_unassign.go
@@ -16,7 +16,7 @@ func newFloatingIPUnassignCommand(cli *CLI) *cobra.Command {
 		Args:                  cobra.ExactArgs(1),
 		TraverseChildren:      true,
 		DisableFlagsInUseLine: true,
-		PreRunE:               cli.ensureActiveContext,
+		PreRunE:               cli.ensureToken,
 		RunE:                  cli.wrap(runFloatingIPUnassign),
 	}
 	return cmd