+++ server_enable_backup.go
@@ -13,7 +13,7 @@ func newServerEnableBackupCommand(cli *CLI) *cobra.Command {
 		Args:                  cobra.ExactArgs(1),
 		TraverseChildren:      true,
 		DisableFlagsInUseLine: true,
-		PreRunE:               cli.ensureActiveContext,
+		PreRunE:               cli.ensureToken,
 		RunE:                  cli.wrap(runServerEnableBackup),
 	}
 	cmd.Flags().String(