+++ sshkey_describe.go
@@ -14,7 +14,7 @@ func newSSHKeyDescribeCommand(cli *CLI) *cobra.Command {
 		Args:                  cobra.ExactArgs(1),
 		TraverseChildren:      true,
 		DisableFlagsInUseLine: true,
-		PreRunE:               cli.ensureActiveContext,
+		PreRunE:               cli.ensureToken,
 		RunE:                  cli.wrap(runSSHKeyDescribe),
 	}
 	return cmd